requirements.txt
bleach==2.1.2
certifi==2017.11.5
chardet==3.0.4
click==6.7
-e git+https://github.com/IQSS/dataverse-client-python.git@0730a657d44782fe8f06d72e9a14e975ac8caa50#egg=dataverse
first==2.0.1
html5lib==1.0.1
idna==2.6
lxml==4.1.1
packaging==16.8
pip-review==1.0
pip-tools==1.11.0
psycopg2==2.7.3.2
pyparsing==2.2.0
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
webencodings==0.5.1
